// src/api/generated.d.ts
export interface components {
    schemas: {
// ...
        ApiResponse_BookmarkState: {
            data: {
                bookmarked: boolean;
                /** Format: uuid */
                topic_id: string;
            };
            meta: components["schemas"]["ResponseMeta"];
        };
// ...
        ApiResponse_PostLikeState: {
            data: {
                /** Format: int64 */
                like_count: number;
                liked: boolean;
                /** Format: uuid */
                post_id: string;
            };
            meta: components["schemas"]["ResponseMeta"];
        };
// ...
        ErrorResponse: {
            error: components["schemas"]["ErrorBody"];
            meta: components["schemas"]["ResponseMeta"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}